# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Sistema de Organización de Eventos — API",
    "version": "1.0.0",
    "description": "API REST del backend del Sistema de Organización de Eventos"
  },
  "servers": [
    {
      "url": "/api",
      "description": "Servidor actual"
    }
  ],
  "components": {
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "token"
      }
    },
    "schemas": {
...
      "Invitado": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "id_evento": {
            "type": "integer"
          },
          "nombre": {
            "type": "string"
          },
          "apellido": {
            "type": "string"
          },
          "email": {
            "type": "string",
            "nullable": true
          },
          "telefono": {
            "type": "string",
            "nullable": true
          },
          "acompanantes_permitidos": {
            "type": "integer"
          },
          "acompanantes_confirmados": {
            "type": "integer",
            "nullable": true
          },
          "notas_especiales": {
            "type": "string",
            "nullable": true
          },
          "estado_confirmacion": {
            "type": "string",
            "enum": [
              "Pendiente",
              "Confirmado",
              "Rechazado"
            ]
          },
          "token_confirmacion": {
            "type": "string"
          },
          "fecha_respuesta": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
    }
  },
  "security": [
    {
      "cookieAuth": []
    }
  ],
  "paths": {
...
    "/eventos/{id}/invitados": {
      "get": {
        "summary": "Lista invitados de un evento",
        "tags": [
          "Invitados"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de invitados",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Invitado"
                  }
                }
              }
            }
          },
          "403": {
            "description": "No autorizado"
          }
        }
      },
      "post": {
        "summary": "Registra un invitado (envía automáticamente el email con el enlace de confirmación)",
        "tags": [
          "Invitados"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "nombre",
                  "apellido"
                ],
                "properties": {
                  "nombre": {
                    "type": "string"
                  },
                  "apellido": {
                    "type": "string"
                  },
                  "email": {
                    "type": "string"
                  },
                  "telefono": {
                    "type": "string"
                  },
                  "acompanantes_permitidos": {
                    "type": "integer"
                  },
                  "notas_especiales": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Invitado creado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Invitado"
                }
              }
            }
          },
          "400": {
            "description": "Faltan campos obligatorios"
          },
          "403": {
            "description": "No autorizado (solo Cliente, y solo sobre su propio evento)"
          }
        }
      }
    },
    "/eventos/{id}/invitados/importar": {
      "post": {
        "summary": "Carga masiva de invitados vía archivo .csv o .xlsx",
        "tags": [
          "Invitados"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "archivo": {
                    "type": "string",
                    "format": "binary",
                    "description": "Columnas esperadas nombre, apellido, email, telefono, acompanantes_permitidos, notas_especiales"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Invitados importados"
          },
          "400": {
            "description": "Archivo faltante o sin filas válidas"
          },
          "403": {
            "description": "No autorizado (solo Cliente)"
          }
        }
      }
    },
    "/invitados/{id}": {
      "put": {
        "summary": "Actualiza datos de un invitado",
        "tags": [
          "Invitados"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Invitado"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Invitado actualizado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Invitado"
                }
              }
            }
          },
          "403": {
            "description": "No autorizado (solo Cliente dueño del evento)"
          },
          "404": {
            "description": "Invitado no encontrado"
          }
        }
      },
      "delete": {
        "summary": "Elimina un invitado",
        "tags": [
          "Invitados"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Invitado eliminado"
          },
          "403": {
            "description": "No autorizado (solo Cliente dueño del evento)"
          },
          "404": {
            "description": "Invitado no encontrado"
          }
        }
      }
    },
    "/invitados/{id}/reenviar-invitacion": {
      "post": {
        "summary": "Reenvía manualmente el enlace de confirmación",
        "tags": [
          "Invitados"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Invitación reenviada",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Invitado"
                }
              }
            }
          },
          "403": {
            "description": "No autorizado (solo Cliente dueño del evento)"
          },
          "404": {
            "description": "Invitado no encontrado"
          }
        }
      }
    },
...
  },
...
}